protocol: pin the sec 5.2 profile region in table consts
